Apollo static analysis - Group C
:Release Notes:
Apollo static analysis - Group C

:Detailed Notes:
http://collab.lge.com/main/display/MMTECH/Apollo+Static+Analysis

CERT ERR30-C
CERT ERR33-C
CERT ARR30-C
CERT STR30-C
CERT STR34-C
CERT CTR50-CPP

:Testing Performed:
Local Build & Test : OK

:QA Notes:
N/A

:Issues Addressed:
[WRP-27521] Apollo static analysis - Group C

+++ include/public/camera/plugin/plugin_interface.hpp
@@ -21,14 +21,14 @@ struct IFeature
 struct IPlugin
 {
     virtual ~IPlugin() {}
-    virtual const char *getName(void)                    = 0;
-    virtual const char *getDescription(void)             = 0;
-    virtual const char *getCategory(void)                = 0;
-    virtual const char *getVersion(void)                 = 0;
-    virtual const char *getOrganization(void)            = 0;
-    virtual size_t getFeatureCount(void)                 = 0;
-    virtual const char *getFeatureName(const int nIndex) = 0;
-    virtual IFeature *createFeature(const char *szName)  = 0;
+    virtual const char *getName(void)                       = 0;
+    virtual const char *getDescription(void)                = 0;
+    virtual const char *getCategory(void)                   = 0;
+    virtual const char *getVersion(void)                    = 0;
+    virtual const char *getOrganization(void)               = 0;
+    virtual size_t getFeatureCount(void)                    = 0;
+    virtual const char *getFeatureName(const size_t nIndex) = 0;
+    virtual IFeature *createFeature(const char *szName)     = 0;
 };
 
 /**
